Remove instant chat in favor for streamed one
