+++ templates/_includes/navbar.html.twig
@@ -1,6 +1,5 @@
 {% set last_conversation = app.session.get('last_conversation') %}
-{% set home_url = last_conversation is defined and last_conversation is not null ? path('chat', {'conversation': last_conversation}) : path('chat') %}
-{% set chat_url = last_conversation is defined and last_conversation is not null ? path('chat_stream', {'conversationId': last_conversation}) : path('chat_stream') %}
+{% set home_url = last_conversation is defined and last_conversation is not null ? path('chat', {'conversationId': last_conversation}) : path('chat') %}
 
 {% set isFirstStart = settings_handler.get.application.hasOpenAIApiKey is same as false %}
 
@@ -26,7 +25,6 @@
             <div class="d-flex flex-column flex-md-row flex-fill align-items-stretch align-items-md-center">
                 <ul class="navbar-nav">
                     <li class="nav-item"><a href="{{ home_url }}" class="nav-link">{{ ux_icon('tabler:message') }}&nbsp;{{ settings_handler.get.chatbotGeneral.chatbotName }}</a></li>
-                    <li class="nav-item"><a href="{{ chat_url }}" class="nav-link">{{ ux_icon('tabler:message') }}&nbsp;{{ settings_handler.get.chatbotGeneral.chatbotName }}<span class="badge bg-red"></span></a></li>
                     <li class="nav-item"><a href="{{ path('library', {'directory': library_root.get.id }) }}" class="nav-link">{{ ux_icon('tabler:books') }}&nbsp;Bibliothek</a></li>
                     <li class="nav-item"><a href="{{ path('world_item_listing') }}" class="nav-link">{{ ux_icon('tabler:database') }}&nbsp;Registratur</a></li>
                     <li class="nav-item"><a href="{{ path('image_generator_overview') }}" class="nav-link">{{ ux_icon('tabler:photo-ai') }}&nbsp;Künste der Mechthild</a></li>
