--- templates/chat/chat.html.twig
@@ -1,57 +1,0 @@
-{% extends 'layout.html.twig' %}
-
-{% block page_title %}{{ conversation.title }}{% endblock %}
-
-{% block page_title_actions %}
-    <div class="col-auto ms-auto d-print-none">
-        <div class="btn-list">
-            <a type="button" class="btn d-none d-sm-inline-block" data-bs-toggle="modal" data-bs-target="#store-conversation-modal">
-                {{ ux_icon('tabler:device-floppy') }}Speichern
-            </a>
-            <a type="button" class="btn btn-icon" title="Einstellungen zum Gespräch" data-bs-toggle="modal" data-bs-target="#conversation-settings">
-                {{ ux_icon('tabler:settings') }}
-            </a>
-            {% if not isTemporary %}
-                <twig:Favorizer:AddToShortcuts id="{{ conversation.id }}" type="{{ constant('class', conversation) }}" asButton="true" />
-            {% endif %}
-            <a type="button" class="btn btn-icon" title="Letzte Gespräche anzeigen" data-bs-toggle="modal" data-bs-target="#latest-conversation-modal">
-                {{ ux_icon('tabler:history') }}
-            </a>
-            <a type="button" class="btn btn-icon" title="Neues Gespräch anfangen" data-bs-toggle="modal" data-bs-target="#create-conversation-modal">
-                {{ ux_icon('tabler:message-plus') }}
-            </a>
-        </div>
-    </div>
-{% endblock %}
-
-{% block page_content %}
-    {% set chatOptions = not isTemporary ? {'conversation': conversation} : {} %}
-
-    <div class="container-xl">
-        {{ component('Chat:Chat', chatOptions) }}
-    </div>
-
-    {% component "Tabler:Modal" with {id: 'store-conversation-modal'} %}
-        {% block modal_full_content %}
-            {{ component('Chat:StoreConversation', { 'conversationId': not isTemporary ? conversation.id : null }) }}
-        {% endblock %}
-    {% endcomponent %}
-
-    {% component "Tabler:Modal" with {id: 'latest-conversation-modal'} %}
-        {% block modal_full_content %}
-            {{ component('Chat:Latest', {entries: 10}) }}
-        {% endblock %}
-    {% endcomponent %}
-
-    {% component "Tabler:Modal" with {id: 'conversation-settings'} %}
-        {% block modal_full_content %}
-            {{ component('Chat:ConversationSettings', chatOptions) }}
-        {% endblock %}
-    {% endcomponent %}
-
-    {% component "Tabler:Modal" with {id: 'create-conversation-modal'} %}
-        {% block modal_full_content %}
-            {{ component('Chat:CreateConversation') }}
-        {% endblock %}
-    {% endcomponent %}
-{% endblock %}
