--- templates/components/chat/latest-conversations.html.twig
@@ -1,32 +1,0 @@
-<div {{ attributes }}>
-    <div class="modal-header">
-        <h3 class="modal-title">Die letzten {{ this.entries }} Gespräche ...</h3>
-        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-    </div>
-
-    <div class="modal-body">
-        {% set conversations = this.conversations %}
-        {% if conversations is empty %}
-            <div class="alert alert-info">Aktuell liegen keine gespeicherten Gespräche vor.</div>
-        {% else %}
-            <table class="table table-vcenter">
-                <thead>
-                    <tr>
-                        <td>Titel</td>
-                        <td class="w-1">&nbsp;</td>
-                    </tr>
-                </thead>
-                <tbody>
-                    {% for entry in this.conversations %}
-                        <tr>
-                            <td>{{ entry.title }}</td>
-                            <td>
-                                <a href="{{ path('chat', {'conversation': entry.id}) }}" title="Öffnen">Öffnen</a>
-                            </td>
-                        </tr>
-                    {% endfor %}
-                </tbody>
-            </table>
-        {% endif %}
-    </div>
-</div>
